### 2019-03-06 #1301 @whedon
[PRE REVIEW]: rrcf: Implementation of the Robust Random Cut Forest algorithm for anomaly detection on streams
**Submitting author:** @mdbartos (<a href="http://orcid.org/0000-0001-6421-222X">Matthew Bartos</a>)
**Repository:** <a href="https://github.com/kLabUM/rrcf" target ="_blank">https://github.com/kLabUM/rrcf</a>
**Version:** v0.2
