+++ paper/paper.bib
@@ -76,3 +76,23 @@
 	title = {A Taxonomy of DDoS Attack and DDoS Defense Mechanisms},
 	journal = {ACM SIGCOMM Computer Communications Review}
 }
+
+@inproceedings{williams_2002_comparative,
+  title={A comparative study of RNN for outlier detection in data mining},
+  author={Williams, Graham and Baxter, Rohan and He, Hongxing and Hawkins, Simon and Gu, Lifang},
+  booktitle={Data Mining, 2002. ICDM 2003. Proceedings. 2002 IEEE International Conference on},
+  pages={709--712},
+  year={2002},
+  organization={IEEE}
+}
+
+@article{tax_2004_support,
+  title={Support vector data description},
+  author={Tax, David MJ and Duin, Robert PW},
+  journal={Machine learning},
+  volume={54},
+  number={1},
+  pages={45--66},
+  year={2004},
+  publisher={Springer}
+}
