### 2019-03-19 #1301 @vc1492a
@VivianePons Thanks, will gladly provide a review of this work. Will read the original paper and review the code, and will comment here (within a few weeks time) with my comments.
