### 2019-03-22 192ef47
JOSS review suggested changes
