+++ paper/paper.bib
@@ -79,6 +79,7 @@
 
 @inproceedings{williams_2002_comparative,
   title={A comparative study of RNN for outlier detection in data mining},
+  doi = {10.1109/ICDM.2002.1184035},
   author={Williams, Graham and Baxter, Rohan and He, Hongxing and Hawkins, Simon and Gu, Lifang},
   booktitle={Data Mining, 2002. ICDM 2003. Proceedings. 2002 IEEE International Conference on},
   pages={709--712},
@@ -88,6 +89,7 @@
 
 @article{tax_2004_support,
   title={Support vector data description},
+  doi = {10.1023/B:MACH.0000008084.60811.49},
   author={Tax, David MJ and Duin, Robert PW},
   journal={Machine learning},
   volume={54},
