+++ rrcf/rrcf.py
@@ -1,5 +1,6 @@
 import numpy as np
 
+
 class RCTree:
     """
     Robust random cut tree data structure as described in:
@@ -54,6 +55,7 @@ class RCTree:
     # Remove point
     >>> tree.forget_point(100)
     """
+
     def __init__(self, X=None, index_labels=None, precision=9):
         # Initialize dict for leaves
         self.leaves = {}
@@ -68,7 +70,8 @@ class RCTree:
                 index_labels = np.arange(X.shape[0], dtype=int)
             self.index_labels = index_labels
             # Check for duplicates
-            U, I, N = np.unique(X, return_inverse=True, return_counts=True, axis=0)
+            U, I, N = np.unique(X, return_inverse=True, return_counts=True,
+                                axis=0)
             # If duplicates exist, take unique elements
             if N.max() > 1:
                 n, d = U.shape
@@ -135,7 +138,7 @@ class RCTree:
         # Determine value for split
         p = np.random.uniform(xmin[q], xmax[q])
         # Determine subset of points to left
-        S1 = (X[:,q] <= p) & (S)
+        S1 = (X[:, q] <= p) & (S)
         # Determine subset of points to right
         S2 = (~S1) & (S)
         # Create new child node
@@ -311,7 +314,7 @@ class RCTree:
         try:
             # Get leaf from leaves dict
             leaf = self.leaves[index]
-        except:
+        except KeyError:
             raise KeyError('Leaf must be a key to self.leaves')
         # If duplicate points exist...
         if leaf.n > 1:
@@ -408,13 +411,14 @@ class RCTree:
             return leaf
         # If leaves already exist in tree, check dimensions of point
         try:
-            assert(point.size == self.ndim)
-        except:
-            raise ValueError("Point must be same dimension as existing points in tree.")
+            assert (point.size == self.ndim)
+        except ValueError:
+            raise ValueError(
+                "Point must be same dimension as existing points in tree.")
         # Check for existing index in leaves dict
         try:
-            assert(not index in self.leaves)
-        except:
+            assert (index not in self.leaves)
+        except KeyError:
             raise KeyError("Index already exists in leaves dict.")
         # Check for duplicate points
         duplicate = self.find_duplicate(point, tolerance=tolerance)
@@ -432,12 +436,12 @@ class RCTree:
             bbox = node.b
             cut_dimension, cut = self._insert_point_cut(point, bbox)
             # TODO: Should this be >= or >?
-            if (cut <= bbox[0, cut_dimension]):
+            if cut <= bbox[0, cut_dimension]:
                 leaf = Leaf(x=point, i=index, d=depth)
                 branch = Branch(q=cut_dimension, p=cut, l=leaf, r=node,
                                 n=(leaf.n + node.n))
                 break
-            elif (cut >= bbox[-1, cut_dimension]):
+            elif cut >= bbox[-1, cut_dimension]:
                 leaf = Leaf(x=point, i=index, d=depth)
                 branch = Branch(q=cut_dimension, p=cut, l=node, r=leaf,
                                 n=(leaf.n + node.n))
@@ -541,8 +545,9 @@ class RCTree:
         if not isinstance(leaf, Leaf):
             try:
                 leaf = self.leaves[leaf]
-            except:
-                raise KeyError('leaf must be a Leaf instance or key to self.leaves')
+            except KeyError:
+                raise KeyError(
+                    'leaf must be a Leaf instance or key to self.leaves')
[…]
