### 2019-03-22 #1336 @vc1492a
@VivianePons My review of this software is provided below.

**Overview**: Overall, I think the authors did a great job implementing the Robust Random Cut Forest algorithm as detailed in the original paper. Below are my comments on particular sections. When the authors strictly meet the requirements laid out in the review checklist, I have marked these items as complete. For those where the requirement is not strictly met, I have left the checklist entry empty. Either way, I have provided some comments below with actions that should be taken (in the case of unchecked entries) and actions that could be taken to improve the work further but that aren't strictly required for submission to JOSS (in the case of checked entries).

I think I will give this implementation a try in a few of my projects! This project is a great implementation of the original paper and could be further improved with a little additional elbow grease. Thoroughly enjoyed reviewing this work!

**License**: while a license is provided in the project repository (MIT), it may be beneficial to indicate the license more publicly in the readme on Github and/or in the documentation. Aa badge could be used or a section added to the readme or documentation.

**Functionality**: I read the original paper in preparation for this review and believe the implementation here in `rrcf` is functionally correct. The software author is correctly accounting for duplicate values and using proper randomization when making cuts to the trees in the forest. The provided examples and unit tests help to verify the functionality of this software package.

However, when reviewing the source code I noticed a few TODO comments in `rrcf.py` that would be better placed in the `dev` branch. It's better practice to keep TODO comments away from the `master` branch. Additionally, there are several areas in `rrcf.py` where any exception that is caught raises a KeyError. This may be misleading, as exceptions which are not in fact due to a KeyError will be raised as KeyErrors and presented as such to users of `rrcf.py`. I noticed a similar pattern with some ValueErrors. I have opened a pull-request that addresses this and a few other minor items, but I will leave this up to the authors to address as it is not an issue with algorithm implementation and rather a software-design decision that should be thought through from a user perspective.

**Version Matching**: The version number submitted to JOSS does match the latest version of the source code in `master` on Github. It may be beneficial to tag the releases that are uploaded to PyPi on Github such that users can download the software directly from Github via a Release as opposed to having to clone or pull the repository.

**Automated Tests**: There are automated unit tests which I ran using `pytest` and `pytest-cov` without issues. It looks like code coverage is currently at 70%. I did not see any explicit instructions on how to run tests for this software. While not explicitly necessary, it may be beneficial to include some instructions on how to run tests for those interested in contributing (maybe this can be included in some contributor's guidelines). Additionally, if you're looking to have others contribute to the project more easily in the future consider setting up a continuous integration pipeline that will automatically run tests and report the test results (e.g. using Travis CI) and coverage (e.g. using Coveralls) with any pushes to `master` or `dev` and for any pull requests opened by open-source contributors. This pipeline would ease the burden on making sure new additions and/or changes to software work as intended, and would also allow maintainers to formally test against every version of Python 3 that should be supported.

**Example Usage**: This software does a good job of providing examples for use, including implementations of the examples from the original paper. The notebooks are also helpful, however I ran into a few issues in the notebooks that I believe should be addressed if they are to continue being included in the repository. First, some datasets are missing from the notebooks that are either 1) not available (nuclear.mat) 2) referenced in the documentation but not on the Github readme. If these datasets could be provided with links to download in the readme that could be beneficial. Second, version control should be addressed - more specifically, I not able to run some of the notebook code with earlier versions of Matplotlib and Pandas but was later to successfully after an update. It would be helpful to new users of the software to have a list of required versions for packages needed to run the examples in the notebooks. Lastly, in the documentation on [the website](https://klabum.github.io/rrcf/taxi.html) the file `../resources/nyc_taxi.csv` is referenced. It may be more clear to simply point to `nyc_taxi.csv` and mention that this file should sit in the directory in which the `rrcf` code is being executed.

**Installation Instructions**:
It is not clear from the Github readme or the provided documentation that numpy is the sole software dependency. It may be beneficial to users to point this out somewhere. However, `install_requires` in `setup.py` is being used properly to ensure numpy is installed if it does not yet exist in the user's environment already. Lastly, the authors should consider having the installation instructions towards the top of the documentation and the readme. This is helpful since the reader would typically start from the top of a document and work their way down when reading documentation and working through examples.

**Community Guidelines**:
Minimal documentation for contributing to the repository exists, indicating that contributors should submit pull requests to the `dev` branch. Developers could improve this by adopting a tagging scheme such as one used in `scikit-learn`. There, pull requests that are a work in progress are tagged [WIP], indicating that someone is working on a fix or new feature and that the work shouldn't be duplicated by someone else. When ready for merging, a two step review process is initiated with the tags [MRG+1] and [MRG+2]. See [contributing pull requests](https://scikit-learn.org/stable/developers/contributing.html) in the `scikit-learn` documentation for more info.

**References**:
Every reference mentioned in the paper is documented as BibTex entries. However, no doi is listed for *A comparative study of RNN for outlier detection in data mining* or *Support Vector Data Description* which have DOIs `10.1109/ICDM.2002.1184035` and `10.1023/B:MACH.0000008084.60811.49`, respectively. There is no listed DOI for *Robust random cut forest based anomaly detection on streams* but I did not locate a matching DOI online for that article. I have included these updates as part of my previously-mentioned pull request.
