### 2019-03-22 #1336 @VivianePons
Thank you very much @vc1492a for this thorough review, @mdbartos I let you work on the comments.
