### 2019-03-23 #1336 @justinshenk
@VivianePons My review of this software is provided below.

**Overview:** I think the authors did a great job implementing the Robust Random Cut Forest algorithm as a Python package. I am also looking forward to implementing this in a few of my projects. Following @vc1492a's thorough review, I have little to add, so my review will rather emphasize the points already addressed which I think are most relevant to the authors of this project.

The code is well commented and described in sufficient detail, which supports future development.

**License:** +1 to adding a badge to the heading.

**Functionality:** Nothing to add.

**Version Matching:** +1 to tagging releases uploaded to GitHub.

**Automated Tests:** Nothing to add.

**Example Usage:** +1 to cleaning up the notebooks if they are to remain.

**Installation Instructions:** +1 to putting installation instructions towards the top of document.

**Community Guidelines:** Nothing to add.

**References:** Nothing to add.
