### 2019-03-23 #1336 @mdbartos
Hi @vc1492a and @justinshenk, thanks for the excellent reviews! I will get to work on these issues as soon as possible.
