### 2019-03-25 38f1c36
JOSS review updates (#48)
* Working version of insert_point

* Fix bug with query and insertpoint

* Cleanup disp and codisp functions

* Reorder methods

* Insertpoint operational, add docstrings

* Small fix

* Remove old files

* Allow empty tree; handle duplicates in insert_point and forget_point

* Account for duplicates in tree construction

* Add ability to print tree

* Update docstrings and minor fixes

* Docstring fix

* Minor fixes

* Bugfix for forget_point

* Add image

* Update readme

* Update README.md

* Type check for point

* Update README.md

* Update README.md

* Store bou
[…]
