### 2019-03-25 4efb995
Move installation instructions; list dependencies
