+++ README.md
@@ -25,6 +25,29 @@ This repository provides an open-source implementation of the RRCF algorithm and
 
 Read the docs [here 📖](https://klabum.github.io/rrcf/).
 
+## Installation
+
+Use `pip` to install `rrcf` via pypi:
+
+```shell
+$ pip install rrcf
+```
+
+Currently, only Python 3 is supported.
+
+### Dependencies
+
+The following dependencies are *required* to install and use `rrcf`:
+
+- [numpy](http://www.numpy.org/)
+
+The following *optional* dependencies are required to run the examples shown in the documentation:
+
+- [pandas](https://pandas.pydata.org/)
+- [scipy](https://www.scipy.org/)
+- [scikit-learn](https://scikit-learn.org/stable/)
+- [matplotlib](https://matplotlib.org/)
+
 ## Robust random cut trees
 
 A robust random cut tree is a binary search tree that can be used to detect outliers in a point set. Points located nearer to the root of the tree are more likely to be outliers.
@@ -208,16 +231,6 @@ for index, point in enumerate(points):
 ```
 
 ![Image](https://github.com/kLabUM/rrcf/blob/master/resources/sine.png)
-
-## Installation
-
-Use `pip` to install:
-
-```shell
-$ pip install rrcf
-```
-
-Currently, only Python 3 is supported.
 
 ## Contributing
 
