### 2019-03-25 0d5453a
Add coveralls support
