+++ .travis.yml
@@ -5,6 +5,10 @@ python:
 install:
   - pip install .
   - pip install -r requirements.txt
+  - pip install pytest-cov
+  - pip install coveralls
 # command to run tests
 script:
-  - pytest
+  - pytest --cov=pysheds/
+after_success:
+  - coveralls
