### 2019-03-26 83203fe
Add classification and comparison to docs
