### 2019-03-26 5e1b7be
Add coveralls badge
