### 2019-03-26 a762b02
Add version numbers to dependencies
