+++ README.md
@@ -39,14 +39,16 @@ Currently, only Python 3 is supported.
 
 The following dependencies are *required* to install and use `rrcf`:
 
-- [numpy](http://www.numpy.org/)
+- [numpy](http://www.numpy.org/) (>= 1.15)
 
 The following *optional* dependencies are required to run the examples shown in the documentation:
 
-- [pandas](https://pandas.pydata.org/)
-- [scipy](https://www.scipy.org/)
-- [scikit-learn](https://scikit-learn.org/stable/)
-- [matplotlib](https://matplotlib.org/)
+- [pandas](https://pandas.pydata.org/) (>= 0.23)
+- [scipy](https://www.scipy.org/) (>= 1.2)
+- [scikit-learn](https://scikit-learn.org/stable/) (>= 0.20)
+- [matplotlib](https://matplotlib.org/) (>= 3.0)
+
+Listed version numbers have been tested and are known to work (this does not necessarily preclude older versions).
 
 ## Robust random cut trees
 
