### 2019-03-26 6b51ca8
Add caveats documentation
