### 2019-03-26 ad83d02
Fix spacing it comparisons documentation
