+++ docs/comparisons.md
@@ -34,30 +34,42 @@ n_inliers = n_samples - n_outliers
 
 # Outlier detectors from sklean plot
 anomaly_algorithms = [
-    ("Robust covariance", EllipticEnvelope(contamination=outliers_fraction)),
-    ("One-Class SVM", svm.OneClassSVM(nu=outliers_fraction,
-                                      kernel="rbf",
-                                      gamma=0.1)),
-    ("Isolation Forest", IsolationForest(contamination=outliers_fraction,
-                                         behaviour='new')),
-    ("Local Outlier Factor", LocalOutlierFactor(n_neighbors=35,
-                                                contamination=outliers_fraction))]
+    ("Robust covariance",
+     EllipticEnvelope(contamination=outliers_fraction)),
+    ("One-Class SVM",
+     svm.OneClassSVM(nu=outliers_fraction,
+                     kernel="rbf",
+                     gamma=0.1)),
+    ("Isolation Forest",
+     IsolationForest(contamination=outliers_fraction,
+                     behaviour='new')),
+    ("Local Outlier Factor",
+     LocalOutlierFactor(n_neighbors=35,
+                        contamination=outliers_fraction))]
 
 # Define datasets
-blobs_params = dict(random_state=0, n_samples=n_inliers, n_features=2)
+blobs_params = dict(random_state=0,
+                    n_samples=n_inliers,
+                    n_features=2)
 datasets = [
-    make_blobs(centers=[[0, 0], [0, 0]], cluster_std=0.5,**blobs_params)[0],
-    make_blobs(centers=[[2, 2], [-2, -2]], cluster_std=[0.5, 0.5],**blobs_params)[0],
-    make_blobs(centers=[[2, 2], [-2, -2]], cluster_std=[1.5, .3],**blobs_params)[0],
-    4. * (make_moons(n_samples=n_samples, noise=.05, random_state=0)[0]
+    make_blobs(centers=[[0, 0], [0, 0]],
+               cluster_std=0.5,**blobs_params)[0],
+    make_blobs(centers=[[2, 2], [-2, -2]],
+               cluster_std=[0.5, 0.5],**blobs_params)[0],
+    make_blobs(centers=[[2, 2], [-2, -2]],
+               cluster_std=[1.5, .3],**blobs_params)[0],
+    4. * (make_moons(n_samples=n_samples,
+                     noise=.05, random_state=0)[0]
                      - np.array([0.5, 0.25])),
-    14. * (np.random.RandomState(42).rand(n_samples, 2) - 0.5)]
+    14. * (np.random.RandomState(42).rand(n_samples, 2)
+           - 0.5)]
 
 # Add outliers to the data sets
 outliers = []  # record keeping
 data = []
 for i in datasets:
-    out = rng.uniform(low=-6, high=6, size=(n_outliers, 2))
+    out = rng.uniform(low=-6, high=6,
+                      size=(n_outliers, 2))
     outliers.append(out)
     data.append(np.concatenate([i, out], axis=0))
 
@@ -75,10 +87,12 @@ for d in range(len(data)):
     tr1 = time.time()
     while len(forest) < num_trees:
         # Select random subsets of points uniformly from point set
-        ixs = np.random.choice(n, size=(n // tree_size, tree_size),
+        ixs = np.random.choice(n,
+                               size=(n // tree_size, tree_size),
                                replace=False)
         # Add sampled trees to forest
-        trees = [rrcf.RCTree(data[d][ix], index_labels=ix) for ix in ixs]
+        trees = [rrcf.RCTree(data[d][ix],
[…]
