### 2019-03-26 e0fe356
Spacing updates to docs
