### 2019-03-26 b4d84ea
Update version post-JOSS review
