+++ setup.py
@@ -3,7 +3,7 @@
 from setuptools import setup
 
 setup(name='rrcf',
-      version='0.2',
+      version='0.3',
       description='Robust random cut forest for anomaly detection',
       author='Matt Bartos, Abhiram Mullapudi, Sara Troutman',
       author_email='mdbartos@umich.edu, abhiramm@umich.edu, stroutm@umich.edu',
