### 2019-03-26 #1336 @mdbartos
Greetings,

@vc1492a and @justinshenk: Thank you again for the extremely relevant and constructive feedback. It would not be an exaggeration to say that this is one of the most helpful reviews I have received on an academic paper :)

A point-by-point summary of changes is listed below:

### License
- We have added the MIT license badge to the README header.

### Functionality
- @vc1492a's pull request has been merged into the master branch. This merge has fixed the error handling behavior. TODO comments have also been removed.

### Version matching
- Historical releases have been tagged.
- A post-review release (0.3) has been tagged and added to pypi.

### Automated tests
- Instructions for running unit tests have been added to the README under *contributing*.
- Additional tests have been added, raising the overall coverage to 84%.
- Tests have been added to verify proper handling of duplicates.
- A coveralls badge has been added to the README to indicate coverage.
- Travis CI and coverall builds are triggered for pull requests and pushes to master. I thought that pull requests to dev were also checked, but I can verify.

### Example usage
- To ease maintainability, notebooks have been folded into the documentation website, and redundant examples have been removed.
- Suggested versions of dependencies needed to run the examples are now included in the README.
- Datasets used in the examples have been added to the /data folder in the main repo, and instructions for using these datasets have been clarified.
- We added a caveats section to the documentation about scaling of dimensions. This addition was posted in reponse to some questions I received about the repo in the past week from another user.

### Installation instructions
- We now list numpy as the sole required dependency in the README.
- In the README, we now also list optional dependencies (and versions) needed to run the examples in the documentation.
- We have moved the installation instructions to the top of the README.

### Community guidelines
- We have expanded the community guidelines by adding suggested types of contributions, guidelines for maintaining code quality and instructions for running unit tests.
- Given that the repo is fairly small and limited in scope (compared to scikit-learn), I have not included a tagging system for pull requests as of yet. However, I am open to implementing one in the future.

### References
- We have merged @vc1492a's pull request, and the requested DOIs are now included. Thanks for the help!

Please let me know if you have any other suggestions. Your time, effort and expertise is very much appreciated.

Thanks again!
MDB
