### 2019-03-26 #1336 @mdbartos
@VivianePons : Would it be possible to change the version to **0.3**? I tried to do this using whedon but was denied.

Also, would it be possible to change the description on JOSS to:

- rrcf is a Python implementation of the Robust Random Cut Forest algorithm: an unsupervised ensemble method for anomaly detection on streaming data

Thank you!
MDB
