### 2019-03-26 #1336 @VivianePons
@arfon do you know where this description is coming from:  https://joss.theoj.org/papers/f8c83c0b01a984d0dbf934939b53c96d ?
