### 2019-03-26 #1336 @arfon
> @arfon do you know where this description is coming from: https://joss.theoj.org/papers/f8c83c0b01a984d0dbf934939b53c96d ?

Yes, it's what the author included in the form when they originally submitted. Once accepted, this text is never public so we can mostly ignore it.
