+++ paper/paper.bib
@@ -19,7 +19,7 @@
 
 @article{breunig_2000_lof,
  author = {Breunig, Markus M. and Kriegel, Hans-Peter and Ng, Raymond T. and Sander, J\"{o}rg},
- title = {LOF: Identifying Density-based Local Outliers},
+ title = {{LOF}: Identifying Density-based Local Outliers},
  journal = {SIGMOD Rec.},
  issue_date = {June 2000},
  volume = {29},
@@ -73,12 +73,12 @@
 	year = 2004,
 	publisher = {ACM},
 	author = {Mirkovic, Jelena and Reiher, Peter},
-	title = {A Taxonomy of DDoS Attack and DDoS Defense Mechanisms},
+	title = {A Taxonomy of {DDoS} Attack and {DDoS} Defense Mechanisms},
 	journal = {ACM SIGCOMM Computer Communications Review}
 }
 
 @inproceedings{williams_2002_comparative,
-  title={A comparative study of RNN for outlier detection in data mining},
+  title={A comparative study of {RNN} for outlier detection in data mining},
   doi = {10.1109/ICDM.2002.1184035},
   author={Williams, Graham and Baxter, Rohan and He, Hongxing and Hawkins, Simon and Gu, Lifang},
   booktitle={Data Mining, 2002. ICDM 2003. Proceedings. 2002 IEEE International Conference on},
