### 2019-03-28 #1336 @justinshenk
> @justinshenk Have you looked at the most recent changes? Do you give your green light for accepting the paper?

@VivianePons Yes, I've reviewed the recent changes from @mdbartos and updated the checklist. Looks good to me!
