### 2019-03-28 #1336 @mdbartos
Hi @VivianePons,

I created an archive with Zenodo using their Github integration service:
https://zenodo.org/record/2613881

Unfortunately, to get the integration to work I had to update the release to 0.3.1

Would it be possible to update the version to 0.3.1 for consistency?

Thanks again,
MDB
