### 2019-03-29 #1336 @VivianePons
Can you update the Zenodo metadata so that the title matches the paper title?

Your name and Sara Troutman's name are also a bit different in both (missing the middle name letter) and the order is different.

Thanks

Viviane
