### 2019-03-29 #1336 @VivianePons
This paper is now good to go @openjournals/joss-eics ! Thank you @mdbartos for the paper and @justinshenk and @vc1492a for the reviews! :) :)
